Move to platformio Full credits to @JyeSmith
