+++ ESP32LapTimer/.gitignore
@@ -1,0 +1,2 @@
+.pio
+.vscode
