+++ ESP32LapTimer/platformio.ini
@@ -1,0 +1,31 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env]
+platform = espressif32
+board = esp32dev
+framework = arduino
+monitor_speed = 115200
+lib_deps =
+    https://github.com/ThingPulse/esp8266-oled-ssd1306
+    https://github.com/adafruit/Adafruit_INA219.git
+
+[env:BOARD_DEFAULT]
+build_flags =
+    -D BOARD=BOARD_DEFAULT
+
+[env:BOARD_TTGO_LORA]
+build_flags =
+    -D BOARD=BOARD_TTGO_LORA
+
+[env:BOARD_OLD]
+build_flags =
+    -D BOARD=BOARD_OLD
+
