+++ ESP32LapTimer/src/ADC.h
