+++ ESP32LapTimer/src/crc.c
